UI 코드 수정 및 에디터 연동

+++ Source/Ch03_ShooterGame/Public/HDPlayerCharacter.h
@@ -24,8 +24,8 @@ public:
 	UPROPERTY(VisibleAnywhere, BlueprintReadOnly, Category = "Camera")
 	UCameraComponent* CameraComp;
 
-	float HP;
-	float MaxHP;
+	int HP;
+	int MaxHP;
 	float Mana;
 	float MaxMana;
